backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0  # Compatible with passlib 1.7.4

# File Upload
python-multipart==0.0.6

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# Data Export
openpyxl==3.1.2

# Supabase Integration
supabase==2.3.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
httpx==0.27.0
requests==2.31.0
# httpx is a dependency of supabase, will be installed automatically

